hardware breakout and test
